Creating flashcards module, simplifying engine loop, and additional cleanup
